ap.go
/* ************************************************************************** */
/*                                                                            */
/*                                                        :::     :::::::::   */
/*   ap.go                                             :+: :+:   :+:    :+:   */
/*                                                   +:+   +:+  +:+    +:+    */
/*   By: atlekbai <tlekbai@doodocs.kz>             +#++:++#++: +#++:++#+      */
/*                                                +#+     +#+ +#+             */
/*   Created: 2024/04/24 13:00:17 by atlekbai    #+#     #+# #+#              */
/*   Updated: 2024/04/24 13:00:16 by atlekbai   ###     ### ###               */
/*                                                                            */
/* ************************************************************************** */

package ap

import (
	"fmt"
	"os"
	"unicode/utf8"
)

// PutRune writes a rune r to stdout. If r is invalid, returns ErrInvalidRune.
func PutRune(r rune) (int, error) {
	if !utf8.ValidRune(r) {
		return 0, ErrInvalidRune
	}

	buf := make([]byte, utf8.RuneLen(r))

	// No need to check return value.
	// Written bytes will be returned from `os.Stdout.Write`
	utf8.EncodeRune(buf, r)

	return os.Stdout.Write(buf)
}

var ErrInvalidRune = fmt.Errorf("invalid rune")
